backend/requirements.txt
# AWS SDK and DynamoDB
boto3>=1.34.0
botocore>=1.34.0

# Authentication and security
PyJWT>=2.8.0
bcrypt>=4.1.0

# Data validation and serialization
pydantic>=2.8.0
email-validator>=2.2.0

# Utilities
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
moto>=4.2.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0